// Web/src/generated/model/releaseWorkflowReviewOptions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReleaseHub API
 * Shared HTTP API contract for the ReleaseHub frontend and backend.
 * OpenAPI spec version: 1.0.0
 */
import type { ReleaseWorkflowReviewRoleOption } from './releaseWorkflowReviewRoleOption';
import type { ReleaseWorkflowReviewUserOption } from './releaseWorkflowReviewUserOption';

export interface ReleaseWorkflowReviewOptions {
  users: ReleaseWorkflowReviewUserOption[];
  roles: ReleaseWorkflowReviewRoleOption[];
}